refactor search modules
