+++ src/command/client/search/event.rs
